pyproject.toml
[project]
name = "streamlit-fit"
version = "0.2.0"
description = "Stock peer analysis dashboard: easily compare stocks against others in their peer group"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=5.5.0",
    "pandas>=2.2.3",
    "streamlit-nightly>=1.44.2.dev20250415",
    "yfinance>=0.2.55",
]
